+++ CHANGELOG.md
@@ -9,3 +9,5 @@
 # `v0.0.11`
 ### Final UI Design and Functionality
 - Parts List population for any selected Process.
+- Model # implication when a Part is selected.
+- UI Entry controls reset their values and states when the Process is changed.